EDITING PROJECT STRUCTURE: Moved database to a seprate folder so it can be shared, only caveat is that you need source setup.sh to get the correct python path of the root project so you can do import database from anywhere
