+++ .gitignore
@@ -1,0 +1,5 @@
+/database/env.py
+__pycache__
+module_pcb_thermal_mockup_v3.5/design-files/module_pcb_thermal_mockup_v3.5/fp-info-cache
+/module_pcb_thermal_mockup_v3.5/design-files/module_pcb_thermal_mockup_v3.5/module_pcb_thermal_mockup_v3.5-backups/
+/module_pcb_thermal_mockup_v3.5/design-files/module_pcb_thermal_mockup_v3.5/~_autosave-module_pcb_thermal_mockup_v3.5.kicad_pcb.lck
