--- control-board/software/soft-v1/.gitignore
@@ -1,2 +1,0 @@
-env.py
-__pycache__/
