+++ database/env.py
@@ -1,0 +1,4 @@
+DATABASE_URI = "postgresql://admin:8euTnKLaCCu@dbod-cms-etl-module-tm-db.cern.ch:6626/admin" 
+#DATABASE_URI = "postgresql://admin:8euTnKLaCCu@localhost:6626/admin" 
+
+#"postgresql://postgres:#MH6vMvjQ!M%gWqd#pn@localhost:5432/tm_mockup_temp"
