--- module_pcb_thermal_mockup_v3.5/software/.gitignore
@@ -1,2 +1,0 @@
-env.py
-__pycache__
